TRIB-3019 start creating parcel and csd packages.

+++ package/common.sh
@@ -41,7 +41,9 @@ function tarFiles()
 {
 	mkdir TESTTAR
 	tar -xvf $1 -C TESTTAR > TAR.LOG
-	rm FILES.LOG CONFIGFILES.LOG
+	if [ -f FILES.LOG ] && [ -f CONFIGFILES.LOG ]; then
+	    rm FILES.LOG CONFIGFILES.LOG
+	fi
 	for path in `cat TAR.LOG`;
 	do
 		fullPath=$path
@@ -161,6 +163,44 @@ function debRules()
 	echo "#export DH_VERBOSE=1"
 	echo "%:"
 	echo -e "\tdh \$@ $RULEOPT"
+}
+
+function package(){
+    packageName=$1
+
+        log "package ${packageName}"
+        config/$packageName/package.sh ${packageName}
+
+}
+
+function createArchive(){
+    packageName=$1
+    log "archive for $packageName dir: $BUILD_DIR"
+    pushd ${BUILD_DIR}
+    if [ $packageName == $PACKAGE_NAME ]; then
+
+        tar zcvf ../$packageName-source.tar.gz . --owner=root --group=root
+        #tar -pczf ../$package-source.tar.gz .
+    fi
+    popd
+}
+
+function cleanBuild(){
+    packageName=$1
+    echo cleanBuildDir
+    echo $packageName
+    if [ -d ${BUILD_DIR}/${packageName} ]; then
+        log "remove old build folder"
+        rm -rf ${BUILD_DIR}/${packageName}
+        else
+        log "no old build folders"
+    fi
+    if [ -f ${BUILD_DIR}/${packageName}-source.tar.gz ]; then
+        log "delete old source file "
+        rm ${BUILD_DIR}/${packageName}-source.tar.gz
+        else
+        log "no old source file"
+    fi
 }
 
 function cleanDeb()
